+++ sdk/pixUser/index.js
@@ -1,0 +1,4 @@
+const pixUser = require('./pixUser.js');
+
+exports.get = pixUser.get;
+exports.PixUser = pixUser.PixUser;